+++ source/kernel/src/drivers/serial_mux.rs
@@ -1,9 +1,7 @@
-use core::ops::{Deref, DerefMut};
-
 use mnemos_alloc::containers::{HeapFixedVec, HeapArc};
 
 use crate::{comms::{
-    bbq::bidi::{BBQBidiHandle, GrantW, GrantR, new_bidi_channel},
+    bbq,
     kchannel::{KChannel, KConsumer, KProducer},
 }, Kernel};
 use maitake::sync::Mutex;
@@ -11,7 +9,7 @@ use maitake::sync::Mutex;
 
 struct PortInfo {
     pub port: u16,
-    pub upstream: BidiProducer,
+    pub upstream: bbq::SpscProducer,
 }
 
 pub struct Message {
@@ -22,8 +20,7 @@ pub struct Message {
 pub enum Request {
     RegisterPort {
         port_id: u16,
-        capacity_in: usize,
-        capacity_out: usize,
+        capacity: usize,
     },
 }
 
@@ -31,13 +28,9 @@ pub enum Response {
     PortRegistered(PortHandle),
 }
 
-struct Outgoing {
-    producer: BidiProducer,
-}
-
 struct Commander {
     cmd: KConsumer<Message>,
-    out: HeapArc<Mutex<Outgoing>>,
+    out: bbq::MpscProducer,
     mux: HeapArc<Mutex<SerialMux>>,
 }
 
@@ -47,10 +40,10 @@ impl Commander {
             let msg = self.cmd.dequeue_async().await.unwrap();
             let Message { req, resp } = msg;
             match req {
-                Request::RegisterPort { port_id, capacity_in, capacity_out } => {
+                Request::RegisterPort { port_id, capacity } => {
                     let res = {
                         let mut mux = self.mux.lock().await;
-                        mux.register_port(port_id, capacity_in, capacity_out).await
+                        mux.register_port(port_id, capacity, &self.out).await
                     };
                     resp.enqueue_async(res.map(|ph| Response::PortRegistered(ph))).await.map_err(drop).unwrap();
                 },
@@ -65,18 +58,26 @@ pub struct SerialMux {
 }
 
 impl SerialMux {
-    async fn register_port(&mut self, port_id: u16, capacity_in: usize, capacity_out: usize) -> Result<PortHandle, ()> {
+    async fn register_port(&mut self, port_id: u16, capacity: usize, outgoing: &bbq::MpscProducer) -> Result<PortHandle, ()> {
         if self.ports.is_full() || self.ports.iter().any(|p| p.port == port_id) {
             return Err(());
         }
+        let (prod, cons) = bbq::new_spsc_channel(self.kernel.heap(), capacity).await;
 
+        self.ports.push(PortInfo { port: port_id, upstream: prod }).map_err(drop)?;
 
-        todo!()
+        let ph = PortHandle {
+            port: port_id,
+            cons,
+            outgoing: outgoing.clone(),
+        };
+
+        Ok(ph)
     }
 }
 
 struct SerialMuxer {
-    incoming: BidiConsumer,
+    incoming: bbq::Consumer,
     mux: HeapArc<Mutex<SerialMux>>,
 }
 
@@ -88,31 +89,39 @@ impl SerialMuxer {
 
 pub struct PortHandle {
     port: u16,
-    cons: BidiConsumer,
-    outgoing: HeapArc<Mutex<Outgoing>>,
+    cons: bbq::Consumer,
+    outgoing: bbq::MpscProducer,
 }
 
-enum Step {
-    Cmd(Message),
-    Inc(GrantR),
-    Out(u16),
+impl PortHandle {
+    pub fn port(&self) -> u16 {
+        self.port
+    }
+
+    pub fn consumer(&self) -> &bbq::Consumer {
+        &self.cons
+    }
+
+    pub fn producer(&self) -> &bbq::MpscProducer {
+        &self.outgoing
+    }
 }
 
 impl SerialMux {
     pub async fn new(
         kernel: &'static Kernel,
         max_ports: usize,
-        serial_port: BBQBidiHandle,
+        serial_port: bbq::BidiHandle,
     ) -> KProducer<Message> {
         let (sprod, scons) = serial_port.split();
+        let sprod = sprod.into_mpmc_producer().await;
 
-        let omutex = kernel.heap().allocate_arc(Mutex::new(Outgoing { producer: sprod })).await;
         let ports = kernel.heap().allocate_fixed_vec(max_ports).await;
-        let imutex = kernel.heap().allocate_arc(Mutex::new(SerialMux { ports })).await;
+        let imutex = kernel.heap().allocate_arc(Mutex::new(SerialMux { ports, kernel })).await;
         let (cmd_prod, cmd_cons) = KChannel::new_async(kernel, max_ports).await.split();
         let commander = Commander {
             cmd: cmd_cons,
-            out: omutex,
+            out: sprod,
             mux: imutex.clone(),
         };
         let muxer = SerialMuxer { incoming: scons, mux: imutex };
